% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{birth2012us}
\alias{birth2012us}
\title{Births in the US in 2012}
\format{A data frame with 990199 rows and 12 columns
  \describe{
    \item{\code{DOB_MM}}{Month of date of birth}
    \item{\code{DOB_WK}}{Day of week of birth}
    \item{\code{MAGER}}{Mother's age}
    \item{\code{TBO_REC}}{Total birth order}
    \item{\code{WTGAIN}}{Weight gain by mother [pounds]}
    \item{\code{SEX}}{a factor with levels \code{F} \code{M}, representing the sex of the child}
    \item{\code{APGAR5}}{APGAR score}
    \item{\code{UPREVIS}}{Number of prenatal visits}
    \item{\code{ESTGEST}}{Estimated weeks of gestation}
    \item{\code{DMETH_REC}}{Delivery Method; levels \code{spontaneous},
    \code{cesarean}, \code{unknown}}
    \item{\code{DPLURAL}}{"Plural Births"; levels \code{single}, \code{twin},
    \code{triplet}, \code{quadruplet}, and \code{quintuplet or higher}}
    \item{\code{DBWT}}{Birth weight [grams]}
  }}
\source{
\url{http://www.cdc.gov/nchs/data_access/Vitalstatsonline.htm}
}
\usage{
birth2012us
}
\description{
A dataset containing information on babies born in the United States during
2012. There is one record per birth. This data set is a random twenty-five percent
sample of the original data set.
}
\keyword{dataset}

