% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirkin.R
\docType{data}
\name{blosum62}
\alias{blosum62}
\title{Amino acid substitution rates}
\format{A data frame with 23 rows and 24 columns}
\source{
Mirkin (2011). Core Concepts in Data Analysis: Summarization,
Correlation and Visualization.
\url{http://www.springer.com/computer/theoretical+computer+science/book/978-0-85729-286-5}
}
\usage{
blosum62
}
\description{
The \strong{BLOSUM62} matrix of substitution scores between amino acids
presented using 1-letter codes.
}
\examples{
blosum62
}
\keyword{datasets}

