% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deck-to-chunk.r
\name{chunkify}
\alias{chunkify}
\title{#Stuff that goes into the slide
#Stuff that goes into the slide}
\usage{
chunkify(chunk)
}
\description{
#Stuff that goes into the slide
#Stuff that goes into the slide
}

