% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-deck.r
\name{clear_cache}
\alias{clear_cache}
\title{Clear chached data in a deckdir}
\usage{
clear_cache(deckdir)
}
\arguments{
\item{deckdir}{Deck directory.}
}
\description{
Clear chached data in a deckdir
}
\keyword{internal}

