% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirkin.R
\docType{data}
\name{company}
\alias{company}
\title{Company}
\format{A data frame with 8 rows and 6 columns:
\describe{
  \item{Company name}{Name of the company (fictional).}
  \item{Income, $mln}{Income in milion dollars.}
  \item{SharP $}{Share price}
  \item{NSup}{The number of principal suppliers.}
  \item{EC}{Yes or No depending on the usage of e-commerce in the company}
  \item{Sector}{which sector of the economy: (a) Retail, (b) Utility, and (c) Industrial}
}}
\source{
Mirkin (2011). Core Concepts in Data Analysis: Summarization,
Correlation and Visualization.
\url{http://www.springer.com/computer/theoretical+computer+science/book/978-0-85729-286-5}
}
\usage{
company
}
\description{
A dataset of eight companies characterized by mixed scale features.
The company names reflect the fact not present in the data - product
affinities: first three companies mostly adhere to product group A,
the next three to product group B, and the last two to product group C.
}
\examples{
company
}
\keyword{datasets}

