% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coursedata-package.r
\docType{package}
\name{coursedata}
\alias{coursedata}
\alias{coursedata-package}
\title{coursedata - Data sets for Introductory/Statistical data analysis with R}
\description{
This coursedata package contains a collection of data sets and helper functions
used for the courses Introductory data analysis with R' and
'Statistical data analysis with R' offered in 2014.
}
\details{
Data sets included in the coursedata package are:

\describe{
   \item{\code{\link{fertility}}}{A dataset containing the bivariate
   distribution of mother's weight and number of offspring for a population of
   rats}
}
}

