% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deck2codechunks}
\alias{deck2codechunks}
\title{Convert rmd decks to code chunk files}
\usage{
deck2codechunks(rmdfile, chunkfile = replace_ext(rmdfile, "R", level = 1),
  header = "", width = 51)
}
\arguments{
\item{rmdfile}{Path to an Rmd file.}

\item{chunkfile}{Path to chunk file.}

\item{header}{Custom text preceding the code chunks.}

\item{width}{Width of the chunk seperators.}
}
\description{
Convert rmd decks to code chunk files
}

