% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fertility}
\alias{fertility}
\title{Rat fertility}
\format{A data frame with 15 rows and 12 columns}
\source{
Pearson 1910
}
\usage{
fertility
}
\description{
A dataset containing the bivariate distribution of mother's weight and number
of offspring for a population of rats.
}
\details{
The \code{rownames} denote the lower bound of the maternal weight in grams;
the \code{colnames} denote the number of offspring. Each number in the table
refers to the number of observations in a particular case, e.g. 38 animals
weighed between 100 and 110 grams and produced 5 offspring.
}
\examples{
fertility
}
\keyword{datasets}

