% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{headache}
\alias{headache}
\title{headache}
\format{A data frame with 30 rows and 6 columns:
\describe{
 \item{\code{subjects}}{index number.}
 \item{\code{sex}}{Sex of the subjects; factor levels \code{F}, \code{M}}
 \item{\code{condition}}{Treatment received by the subject; factor levels
   \code{placebo}, \code{aspirin}}
 \item{\code{before}}{Headache intensity before treatment.}
 \item{\code{after}}{Headache intensity after treatment.}
 \item{\code{change}}{Change in headache intensity.}
}}
\usage{
headache
}
\description{
A fictional dataset containing the effects of asprin and a placebo treatment
of female and male patients
}
\keyword{datasets}

