% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hiv}
\alias{hiv}
\title{Gene expression data HIV}
\format{A data frame with 216000 rows and 5 columns:
\describe{
  \item{\code{sample}}{Sample code (6 samples per \code{tissue} per \code{state})}
  \item{\code{state}}{\code{control} or \code{HIV}}
  \item{\code{tissue}}{One of \code{Basal ganglia}, \code{Frontal cortex},
    or \code{White matter}}
  \item{\code{gene}}{One of 6000 gene objects}
  \item{\code{expr}}{log-transformed expression values}
}}
\source{
Gelman BB, Chen T, Lisinicchia JG, Soukup VM et al.
The National NeuroAIDS Tissue Consortium brain gene array: two types of
HIV-associated neurocognitive impairment. PLoS One 2012;7(9):e46178.
PMID: 23049970
}
\usage{
hiv
}
\description{
A random sample of 6000 genes from gene expression data set GDS4358.
Contains expression data from basal ganglia, frontal cortex, and white matter
tissue from HIV patients and controls.
}
\examples{
summary(hiv)
}
\keyword{datasets}

