% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialise-deck.r
\name{initialise_deck}
\alias{initialise_deck}
\title{Intitialise (or open) an Rmd deck.}
\usage{
initialise_deck(deckdir, title = "", subtitle = "", author = "",
  job = "", libs = "", slides = "## First Slide\\n\\n---\\n\\n",
  open_rmd = TRUE, replace = FALSE)
}
\arguments{
\item{deckdir}{Directory in which the Rmd template will be created (or
in which an \emph{index.Rmd} file is present)}

\item{title}{Title string to be substituted into the template.}

\item{subtitle}{Subtitle string to be substituted into the template.}

\item{author}{Author name to be substituted into the template. Taken from
the DESCRIPTION file if not provided.}

\item{job}{Job or address to be substituted into the template.}

\item{slides}{Slides to be substituted into the template.}

\item{open_rmd}{Open the index file for edition after it has been created.}
}
\description{
This function creates a deck directory and initialises an Rmd slidify
template for the user to edit.
}

