% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialise-vignette.r
\name{initialise_vignette}
\alias{initialise_vignette}
\title{Intitialise (onr open) an Rnw vignette.}
\usage{
initialise_vignette(pkg = ".", title = "My Vignette",
  short_title = gsub("\\\\s+", "-", title), address = "", open_rnw = TRUE)
}
\arguments{
\item{pkg}{Root directory of the package for which the vignette is going to
be created. Must contain a valid DESCRIPTION file}

\item{title}{Title of the vignette.}

\item{short_title}{File name of the vignette.}

\item{address}{The authors address.}

\item{open_rnw}{Whether to open the Rnw file created for editing.}
}
\description{
This function creates a 'vignettes' directory and initialises an Rnw vignette
template for the user to edit.
}

