% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirkin.R
\docType{data}
\name{intrusion}
\alias{intrusion}
\title{Intrusion}
\format{A data frame with 100 rows and 7 columns:
\describe{
  \item{Pr}{The protocol-type, which is either tcp or icmp or udp.}
  \item{BySD}{The number of data bytes from source to destination.}
  \item{SH}{The number of connections to the same host as the current one in the past 2 s.}
  \item{SS}{The number of connections to the same service as the current one in the past 2 s.}
  \item{SE}{The rate of connections (per cent in SHCo) that have SYN errors.}
  \item{RE}{The rate of connections (per cent in SHCo) that have REJ errors.}
  \item{A}{The type of attack (ap - apache, sa - saint, sm - smurf, and no attack)}
}}
\source{
Mirkin (2011). Core Concepts in Data Analysis: Summarization,
Correlation and Visualization.
\url{http://www.springer.com/computer/theoretical+computer+science/book/978-0-85729-286-5}
}
\usage{
intrusion
}
\description{
A dataset consisting of a hundred communication packages along with some of
their features sampled from a set of artificially created data publicly
available on webpage of MIT Lincoln Laboratory:
\url{http://www. ll.mit.edu/mission/communications/ist/corpora/ideval/data/intex.html}.
}
\examples{
intrusion
}
\keyword{datasets}

