% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirkin.R
\docType{data}
\name{market_towns}
\alias{market_towns}
\title{Market towns}
\format{A data frame with 45 rows and 13 columns:
\describe{
  \item{Town}{Name of the town.}
  \item{Pop}{Population resident.}
  \item{PS}{Primary schools.}
  \item{D}{General Practitioners.}
  \item{Hos}{Hospitals.}
  \item{Ba}{Banks.}
  \item{Sst}{Superstores.}
  \item{Pet}{Petrol stations.}
  \item{DIY}{Do It Yourself shops.}
  \item{Swi}{Swimming pools.}
  \item{Po}{Post offices.}
  \item{CAB}{Citizen Advice Bureaus.}
  \item{FM}{Farmer markets.}
}}
\source{
Mirkin (2011). Core Concepts in Data Analysis: Summarization,
Correlation and Visualization.
\url{http://www.springer.com/computer/theoretical+computer+science/book/978-0-85729-286-5}
}
\usage{
market_towns
}
\description{
A dataset of Market towns in West Country, England is presented along
with features characterizing population and social infrastructure according
to a census in 1991.
}
\examples{
market_towns
}
\keyword{datasets}

