% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyt}
\alias{nyt}
\title{New York Times ads}
\format{A data frame with 458441 rows and 5 columns:
\describe{
  \item{\code{Age}}{ Age of user. }
  \item{\code{Gender}}{ Gender of user (0=female, 1=male). }
  \item{\code{Impressions}}{ The number of times an ad is seen. Clicking or
  not is not taken into account. }
  \item{\code{Clicks}}{ The number of times an ad has been clicked at. }
  \item{\code{Signed_In}}{ Has the user been signed in (0=no, 1=yes). }
}}
\source{
\url{https://github.com/oreillymedia/doing_data_science}
}
\usage{
nyt
}
\description{
This data set represents a (simulated) day’s worth of ads shown and
clicks recorded on the New York Times home page in May 2012. Each
row represents a single user.
}
\keyword{datasets}

