% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{primates}
\alias{primates}
\title{Primate taxonomy}
\format{A data frame with 22 rows and 12 columns
  \describe{
    \item{\code{Suborder}}{Suborder (Parvorder) classification: \code{Strepsirrhini}
    (lemurs, lorises, etc.), \code{ Platyrrhini} (New World monkeys), \code{Catarrhini}
    (Old World anthropoids)}
    \item{\code{Family}}{Family classification}
    \item{\code{Species}}{Species classification}
    \item{\code{PregLen}}{Average duration of pregnancy [days]}
    \item{\code{AgeAtBirth}}{Earliest age at birth [years]}
    \item{\code{AgeAtDeath}}{Maximum age at death [years]}
    \item{\code{BodyWeightBirth}}{Body weight at birth [kg]}
    \item{\code{MaternalBodyWeigthAtBirth}}{Maternal body weight at birth [kg]}
    \item{\code{BrainWeightBirth}}{Brain weight at birth [g]}
    \item{\code{MaternalBrainWeigthAtBirth}}{Maternal brain weight at birth [g]}
    \item{\code{SO}}{Suborder code}
    \item{\code{FA}}{Family code}
  }}
\source{
Pruscha H. (2006) Statistisches Methodenbuch. Springer Berlin.
ISBN-13 978-3-540-26006-6
}
\usage{
primates
}
\description{
A dataset containing information on 22 primate species on variables like
maximum age at death and several birth and pregnancy-related data.
}
\keyword{dataset}

