% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{savings}
\alias{savings}
\title{Savings rates in 50 countries}
\format{A data frame with 50 rows and 5 columns:
\describe{
 \item{\code{sr}}{savings rate - personal saving divided by disposable income}
 \item{\code{pop15}}{percent population under age of 15}
 \item{\code{pop75}}{percent population over age of 75}
 \item{\code{dpi}}{per-capita disposable income in dollars}
 \item{\code{ddpi}}{percent growth rate of dpi}
}}
\source{
Belsley, D., Kuh. E. and Welsch, R. (1980) "Regression Diagnostics"
Wiley.
}
\usage{
savings
}
\description{
The \code{savings} data frame has 50 rows and 5 columns. The data is
averaged over the period 1960-1970.
}
\details{
Also appears as \code{LifeCycleSavings} in the \code{datasets} package,
and as \code{savings} in the \code{faraway} package.
}
\keyword{datasets}

