% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{spodoptera}
\alias{spodoptera}
\title{\emph{Spodoptera frugiperda} mating trials}
\format{A data frame with 3129 rows and 18 columns
  \describe{
    \item{\code{Cross}}{Type of cross (Backcross, F1, or Parental)}
    \item{\code{Female}}{Strain of the female mating partner (R, C, RC, CR),
    where CR and RC describe female x male hybrids.}
    \item{\code{MGM}}{Strain of the maternal grandmother}
    \item{\code{MGF}}{Strain of the maternal grandfather}
    \item{\code{Male}}{Strain of the male mating partner.}
    \item{\code{PGM}}{Strain of the paternal grandmother}
    \item{\code{PGF}}{Strain of the paternal grandfather}
    \item{\code{Day}}{Day of experiment}
    \item{\code{Pair}}{Number of mating pair (= number of mating cup)}
    \item{\code{NumCop}}{Number of copulation [First, Second, Third, ...]}
    \item{\code{NumFcal}}{Number of female calling [First, Second, Third, ...]}
    \item{\code{NumOvi}}{Number of oviposition [First, Second, Third, ...]}
    \item{\code{NumMcal}}{Number of male calling [First, Second, Third, ...]}
    \item{\code{StartCop}}{Start of copulation [hours into scotophase]}
    \item{\code{DurationCop}}{Duration of copulation [hours]}
    \item{\code{StartFcal}}{Start of female calling [hours into scotophase]}
    \item{\code{StartOvi}}{Start of oviposition [hours into scotophase]}
    \item{\code{StartMcal}}{Start of male calling [hours into scotophase]}
  }}
\source{
\url{http://onlinelibrary.wiley.com/doi/10.1111/j.1420-9101.2009.01759.x/abstract}
}
\usage{
spodoptera
}
\description{
The noctuid moth \emph{Spodoptera frugiperda} (fall armyworm) consists of two
strains (R and C) associated with different larval host plants (most notably
corn and rice). These strains exhibit differential temporal patterns of reproductive
behaviours (copulation, female calling, male calling, oviposition) during scotophase,
with the corn strain more active earlier in the night. This data set describes
an experiment where the reproductive timing of 320 fall armyworm mating
pairs was observed for three consecutive nights.
}
\keyword{datasets}

