% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirkin.R
\docType{data}
\name{student}
\alias{student}
\title{Student}
\format{A data frame with 100 rows and 6 columns:
\describe{
  \item{Oc}{Occupation. Either Information technology (IT)
    or Business Administration (BA) or anything else (AN)}
  \item{Age}{Age, in years.}
  \item{Ch}{Number of children.}
  \item{SE}{Marks in Software and Programming.}
  \item{OO}{Marks in Object-Oriented Programming.}
  \item{CI}{Marks in Computational Intelligence.}
}}
\source{
Mirkin (2011). Core Concepts in Data Analysis: Summarization,
Correlation and Visualization.
\url{http://www.springer.com/computer/theoretical+computer+science/book/978-0-85729-286-5}
}
\usage{
student
}
\description{
A fictitional dataset is presented as imitating a typical set up for a group
of Birkbeck University of London part-time students pursuing Master’s degree in
Computer Sciences.
}
\details{
This dataset refers to a hundred students along with six features, three of which
are personal characteristics.
}
\examples{
student
}
\keyword{datasets}

