% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{teengamb}
\alias{teengamb}
\title{Study of teenage gambling in Britain}
\format{A data frame with 47 rows and 5 columns:
\describe{
  \item{\code{sex}}{ 0=male, 1=female. }
  \item{\code{status}}{ Socioeconomic status score based on parents' occupation. }
  \item{\code{income}}{ In pounds per week. }
  \item{\code{verbal}}{ Verbal score in words out of 12 correctly defined. }
  \item{\code{gamble}}{ Expenditure on gambling in pounds per year. }
}}
\source{
Ide-Smith & Lea, 1988, Journal of Gambling Behavior, 4, 110-118
}
\usage{
teengamb
}
\description{
The \code{teengamb} data frame contains survey data from a study
conducted to study teenage gambling in Britain.
}
\keyword{datasets}

