% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{uswages}
\alias{uswages}
\title{Weekly wages of US male workers in 1988}
\format{A data frame with 2000 rows and 10 columns:
\describe{
  \item{\code{wage}}{ Real weekly wages in dollars (deflated by personal
  consumption expenditures - 1992 base year) }
  \item{\code{educ}}{ Years of education }
  \item{\code{exper}}{ Years of experience }
  \item{\code{race}}{ 1 if Black, 0 if White (other races not in sample) }
  \item{\code{smsa}}{ 1 if living in Standard Metropolitan Statistical Area,
  0 if not }
  \item{\code{ne}}{ 1 if living in the North East }
  \item{\code{mw}}{ 1 if living in the Midwest }
  \item{\code{we}}{ 1 if living in the West }
  \item{\code{so}}{ 1 if living in the South }
  \item{\code{pt}}{ 1 if working part time, 0 if not }
}}
\source{
Bierens, H.J., and D. Ginther (2001): "Integrated Conditional Moment
Testing of Quantile Regression Models", Empirical Economics 26, 307-324
}
\usage{
uswages
}
\description{
The \code{uswages} data frame contains data on weekly Wages for
US male workers sampled from the Current Population Survey in 1988.
}
\keyword{datasets}

