% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{birth2012us}
\alias{birth2012us}
\title{Births in the US in 2012}
\format{A data frame with 990199 rows and 12 columns
\describe{
 \item{DOB_MM}{Month of date of birth}
 \item{DOB_WK}{Day of week of birth}
 \item{MAGER}{Mother's age}
 \item{TBO_REC}{Total birth order}
 \item{WTGAIN}{Weight gain by mother [pounds]}
 \item{SEX}{a factor with levels \code{F} \code{M}, representing the sex of the child}
 \item{APGAR5}{APGAR score}
 \item{UPREVIS}{Number of prenatal visits}
 \item{ESTGEST}{Estimated weeks of gestation}
 \item{DMETH_REC}{Delivery Method; levels \code{spontaneous}, \code{cesarean},
   \code{unknown}}
 \item{DPLURAL}{"Plural Births"; levels \code{single}, \code{twin}, \code{triplet},
   \code{quadruplet}, and \code{quintuplet or higher}}
 \item{DBWT}{Birth weight [grams]}
}}
\source{
\url{http://www.cdc.gov/nchs/data_access/Vitalstatsonline.htm}
}
\usage{
birth2012us
}
\description{
A dataset containing information on babies born in the United States during
2012. There is one record per birth. This data set is a random twenty-five percent
sample of the original data set.
}
\keyword{datasets}

