% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dax30}
\alias{dax30}
\title{Dax30}
\format{A data frame with 7818 rows and 8 columns
\describe{
 \item{symbol}{A factor representing the date for each quote}
 \item{Date}{Date on which the quote was taken}
 \item{Open}{A numeric vector representing the opening price for each stock on each day}
 \item{High}{A numeric vector representing the high price for each stock on each day}
 \item{Low}{A numeric vector representing the low price for each stock on each day}
 \item{Close}{A numeric vector representing the closing price for each stock on each day}
 \item{Volume}{A numeric vector representing the number of shares traded for each stock on each day}
 \item{Adj.Close}{A numeric vector representing the closing price for each stock on each day}
}}
\source{
\url{http://de.finance.yahoo.com/}
}
\usage{
dax30
}
\description{
A dataset containing the DAX 30 stock quotes between 27. Feb. 2013 and
26. Feb. 2014.
}
\keyword{datasets}

