% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-deck.r
\name{load_deck}
\alias{load_deck}
\title{Load a html5 deck}
\usage{
load_deck(pkg = "IDA")
}
\arguments{
\item{pkg}{Package from which to load slides.}
}
\description{
This function lets you choose among html5 presentations available
in the package \code{pkg}, and loads it into your browser.
}

